+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - Enums in JSON schema are type sensitive, so a warning will be issued for types that have capital letters. To fix these warnings, change your type names in your `.ysls.json` file to be lowercase. (These warnings have no impact on your Yarn script editing experience or runtime behaviour.)
 - Empty nodes will no longer be included in the compiled output
   - a warning diagnostic will be generated for each empty node
+- Fixed a bug where set-referencing inferred value set statements would crash the compiler
 
 ### Removed